Only pull the BLOB when needed. Better error.
This turned out to be fun. Google found me an old email from 2010
<https://www.mail-archive.com/dbix-class@lists.scsys.co.uk/msg04161.html>.
A message there indicated fREW Schmidt is planning on implementing.
Found out he did, at
<https://metacpan.org/pod/DBIx::Class::Helper::ResultSet::RemoveColumns>
and that even better, there is one to do it automatically
<https://metacpan.org/pod/DBIx::Class::Helper::ResultSet::AutoRemoveColumns>.

Implemented that, and had the JPEG controller explicitly ask for the
column... and it was ignored. After some fiddling (do I remember DBIC
syntax right? Yes.) turns out to be a bug in that RemoveColumns
module... Bug reported, along with a quick patch
<https://github.com/frioux/DBIx-Class-Helpers/issues/27> and it now
works.

Finally.

Also, when an image isn't found, don't die(). Instead, send back a 404.