Simplify the contacts data model
Previously, had contact info for each passholder, as well as emergency
contacts for each passholder.

Switch to having contacts for an entire family. Each contact can be an
administrative contact and/or an emergency contact.

This will make the data entry form much simpler. It was looking to be a
nightmare.