Edit passholder form mostly done.
Big thing left to do is plug in the video capture to take the photo. Got
the docs in front of me, have it sort of working (before even reading
the docs...). Only works in Chrome so far, Firefox does not like it
(pegs a CPU, does nothing, only restarting browser unpegs CPU).