New design, without non-rectangular text regions.
Not sure how to do a non-rectangular text region in Pango. The easy
layout API doesn't appear to do it. Maybe you can ask it for lines, and
change the width... not sure. Easier to just change the card design.

Also, there is supposed to be white on both edges of the bar code. Add
it.