Switch to database sessions.
For some reason, MMAP sessions are not working right on the pool laptop.
Already have a running Pg instance, so might as well use that. Use an
unlogged table because its OK to lose session data on a crash. It just
causes a logout.

Could also do memcached, but that'd add more dependencies.