Database changes for documents feature.
Basic plan is to store documents (something like "Pool Rules"), and the
versions of each (identified just by a date). Also store with the
document which ages need to agree to/be shown it. Finally, store (for
each passholder) which documents & versions he/she has agreed to/been
shown.

In all likelyhood, there will be just one document... but probably not
much more work to support N documents, so going ahead and doing so.

Working towards issue #9.

As to cascades: If a passholder is deleted, we no longer need or want to
know which documents he/she had seen. Same if we delete a document—we no
longer care about its versions or who had seen it. Same with a version,
if we delete a version, doesn't matter who has seen it.

Note: I don't expect to be deleting documents or versions except very
rarely.