Needed documents on the checkin page.
If there are needed documents, they show up on the checkin page in a
blue box together with a message asking the guard to present them. Each
document has a radio group (yes/no) about whether it was presented. If
set to yes, then when checked in, update the passholder_documents table.

Further work towards issue #9. All that's left is when adding a
passholder (admin side).