Report backup failure due to out of space.
IPC::Run3 doesn't die for all errors. In particular, if the child
process is run but returns a non-zero exit code, you have to check $? to
detect. So check $?.

Also improved the visibility of the error and added the exit code to the
output.

Closes #18. Also checked on the log export—that also notices, but
doesn't handle as gracefully. You'll get a Catalyst error (come back
later page). Good enough for now.